requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
python-dateutil>=2.8.2
Pillow>=10.0.0
psycopg2-binary>=2.9.9
Pillow==11.2.1
python-decouple==3.8
django-widget-tweaks==1.5.0
